requirements.txt
numpy
pillow==12.0.0
python-dotenv==1.2.1
requests~=2.32.5
colorlog~=6.10.1
colour-science~=0.4.6
networkx==3.5